MASSIVE rework of the Point system to allow for floating point precision.

+++ src/FE_SRC_CLIENT/com/ForgeEssentials/client/network/PacketSelectionUpdate.java
@@ -35,12 +35,12 @@ public class PacketSelectionUpdate implements IForgeEssentialsPacket
 	@SideOnly(Side.CLIENT)
 	public static void readClient(DataInputStream stream, WorldClient world, EntityPlayer player) throws IOException
 	{
-		// point 1 available.
+		// podouble 1 available.
 		if (stream.readBoolean())
 		{
-			int x = stream.readInt();
-			int y = stream.readInt();
-			int z = stream.readInt();
+			double x = stream.readDouble();
+			double y = stream.readDouble();
+			double z = stream.readDouble();
 
 			ForgeEssentialsClient.getInfo().setPoint1(new Point(x, y, z));
 		}
@@ -49,12 +49,12 @@ public class PacketSelectionUpdate implements IForgeEssentialsPacket
 			ForgeEssentialsClient.getInfo().setPoint1(null);
 		}
 
-		// point 2 available
+		// podouble 2 available
 		if (stream.readBoolean())
 		{
-			int x = stream.readInt();
-			int y = stream.readInt();
-			int z = stream.readInt();
+			double x = stream.readDouble();
+			double y = stream.readDouble();
+			double z = stream.readDouble();
 
 			ForgeEssentialsClient.getInfo().setPoint2(new Point(x, y, z));
 		}
